BS: build uglifyjs-cs
UglifyJS was meant to be used on Node.js but with really small
changes it can be used in the browser(or in other Js platforms).
However UJS is in active development so if we want to keep
updated we should have to re-apply these changes over and over again.
This commit address this issue.
The original unmodified UJS files are assembled together with
some glue code that 'fakes' the presence of Node.js module system.