+digg
Added digg support. Since digg doesn't record the date you dugg a link
(reference: http://twitter.com/#!/diggapi/status/116276946203197441),
it will take the date the link got promoted to the frontpage (for diggs
and comments). Although not correct, it should be sufficient for most
users.

For submissions, the correct take is available and used.