+wikipedia (English)
Added support for wikipedia contributions, english language only.
[Wikipedia Api Reference](http://www.mediawiki.org/wiki/API:Query)

By default 10 contributions are returned, but that number can be
configured. Thinking of setting it to the lifestream count.

The API doesn't return a url for the contribution. It would be possible
to get it, but that would require another API call, so I chose to
simply replace spaces with underscores in the title, since this is the
way wikipedia generates urls anyway (afaik).

The username used in the example has the highest number of edits on
wikipedia.

Maybe should add a setting for the chosen language …