Add support for FollowEvent
There is a slight issue with the FollowEvent. It doesn't have a
payload.action value like many of the other events do. The
payload.action value is supposed to contain a verb that describes the
performed action such as "opened", "closed", "started", "stopped", etc.

Now this isn't *curently* a problem as the only type of FollowEvents
that show up in the feeds are of the "started" type so the word
"started" can just be hard-coded to replace the missing payload.action
value.

Of course if things change in the future this will need to be changed.