FinalGradesTranslator was returning a nil source.
RegStatusTranslator was returning "Bearfacts:RegStatus" as the source.

Both classes should be returning the string "Bear Facts" as the value for :source.

Changed both classes and their respective tests.

Also had to change a model test that checked for Bearfacts:RegStatus in the source field of the notifications. That literal is no longer there, so now it's just checking for Bear Facts notifications in general

Changed a model test that checked for Bearfacts:RegStatus in the source field of the notifications. That literal is no longer there, so now it's just checking for Bear Facts notifications