CLC-notask-debugger-gem
Require ruby-debug gem in an oddly named group ":shell_debug".
The reason for the group is to have the gem loaded without forcing
it into runtime, which would cause problems for developers using
RubyMine, like Ray.

With this change,
- the rails server can be started in debug mode with the --debugger flag
  > rails s --debugger
- the debugger can be run with rspec by using the --debug flag
  > RAILS_ENV=test rspec spec --debug

It's kind of lame that the flag name is different for rails s and rspec,
but it works.  The important thing is that the ruby debugger can now be
enjoyed by RubyMine IDE users and regular old command-line folk, like me.
Thanks Ray, for your help in finding a solution that works for everyone!

group :shell_debug do
  gem 'ruby-debug', '>= 0.10.5.rc9'
end