CLC-3005: Can't edit links or superusers in ccadmin
https://jira.media.berkeley.edu/jira/browse/CLC-3005

The rails_admin gem became out of sync with Font Awesome 4,
which we recently updated in the project. For any list page
the action (edit/delete) icons were there, but the css prefix
was wrong.

Unfortunately, an upgrade of rails_admin was not possible since
it depends on Rails 4, so we had to do a workaround here, renaming
the rails_admin theming.css to theming.scss in the
app/assets/stylesheets/rails_admin/custom/ directory and
adding the lines found in the rails_admin 0.6 codebase for fixing
this problem:

app / assets / stylesheets / rails_admin / base / font-awesome-4-compability.scss