CLC-3090: FinAid Messages: Dev environment is only showing old messages (mostly from 2012)
https://jira.media.berkeley.edu/jira/browse/CLC-3090

Changed the default settings for myfinaid_proxy.term_year to 2014

Important to note that this configuration setting is a sitting duck
for a future bug as it will have to be changed at just the right time
for the upcoming academic year unless we can find a way to stop
using it before then.  See Mike's comments to Bernie in this Jira.

Also, Settings.myfinaid_proxy.base_url is still pointing to
myfinaid's development api "https://apis-dev.berkeley.edu/myfinaid"