CLC-2820: Update bundler and use its retry option
https://jira.media.berkeley.edu/jira/browse/CLC-2820

Change in codebase
Added the --retry to bundle install --local with a value of 3.
-bundle install --local || ...
+bundle install --local --retry 3 ||  ...

DevOps changes on ets-dev. I did the following:
- Set default ruby to jruby-1.7.8
- Removed jruby-1.7.6 and all associated gemsets
- Updated bundler to version 1.5.2 for global and for all named gemsets (calcentral, calcentral_qa, chris, mike, ray, jason, christian)

Monitoring:
On ets-dev, under my plan related to this branch, using the "mike" gemset,
I uninstalled several test gems (spork, capybara, simpleconv), hoping to
see some network errors in the bamboo logs that would force the retries
to occur.

After several rebuilds, I was unable to actually see the retries in the
bamboo logs, but nothing broke either, so I'm going to send up this PR
after having checked all this with Chris.

If we start seeing network errors on future bamboo builds, we can check
the logs to see if the retries are actually being performed and reopen
this ticket, if necessary.