CLC-3125-deep-linking - Third try
Pass :url to /auth/cas in an attempt to give the CAS server what it
expects in order to have it returned to our callback method
/auth/cas/callback.

Previous attempts have been made naming the parameter :smart_path
and also :service, but although this works on local dev machines, it
can't be trusted.  CAS gives "special behavior" to localhost, so what
we see working before the PR can end up breaking when it gets to dev.