CLC-3272-missing-year-in-finaid-feed
https://jira.ets.berkeley.edu/jira/browse/CLC-3272

For certain UID's, that seem to be long to registered test students
the finaid proxy returns an xml document with no data and a non-sucessful
response code such as:

<SSIDOC>
<Response>
  <Code>B0023</Code>
  <Message>FAILED - BIO record does not exist</Message>
</Response>
</SSIDOC>

On success, the Response Code comes back as 00000.

So, the fix here is to check those response codes and log a warning
with the uid, code, and message returned.

Previously, this was logged as an error saying that the user couldn't
access the document's AidYear, which was misleading because the
document simply had no data.