CLC-3463: Factor student info out of the rootscope https://jira.ets.berkeley.edu/jira/browse/CLC-3463
I've been going back and forward for a while between this solution and a more complex one.

This should be straightforward to understand and imo be a good foundation to build on.

The thing we'll need to test for is looking whether we are making too many requests to the same endpoint.