Users who can act-as should be able to act-as any user
https://jira.ets.berkeley.edu/jira/browse/CLC-3419

Reproduced the error:
Made an integration test that fails in order to reproduce the bug:
"make sure admin users can act as a user who has never signed in before"
even in production
Thanks for your help, Chris!

The fix:
Removed a block of code from the SessionsController that didn't allow
viewing of users that had never logged in production.