CLC-3538 - Error from ETSBlog
https://jira.ets.berkeley.edu/jira/browse/CLC-3538

Made the alert teaser field optional.
Only include :teaser in the result hash if it is present in the xml document.
Added dummy json file for it so Christian can test it on the splash page.
If it turns out it's necessary for the splash page, we can open another
jira and I'll make that check