CLC-3432: User alert feature can't handle diacriticals
https://jira.ets.berkeley.edu/jira/browse/CLC-3432

Added "# encoding: utf-8" to the top of my specs so they'd pass
and added it to the top of the blog_feed_controller so the feed
will return the correct utf-8 characters.

Made a fake xml file that includes all kinds of diacritical characters, even in arabic and greek.
Made failing test to go along with it.