CLC-3594: NameError in MyActivitiesController#get_feed
https://jira.ets.berkeley.edu/jira/browse/CLC-3594

Separated the class Canvas from the module MyActivities and renamed
the class CanvasActvities to avoid confusion with the modlule named
canvas.

As Chris pointed out, this was happening because
app/models/my_activities/canvas.rb had the same name
as Canvas the module.

Confirmed that the NameError in MyActivitiesController#get_feed
no longer appears.