Remove large size for GPA
Most students visit the Academics tab to check things other than GPA,
so there's no particular reason for GPA to be larger.

In addition, many students visit CalCentral in public settings and
would prefer to have their GPA less visible to people around them.

This is a temporary solution, but a "show GPA" checkbox might be more
appropriate in the future.